requirements.txt
fastapi
uvicorn[standard]
libcst
asyncpg
sqlalchemy[asyncio]
alembic
python-dotenv
pydantic[email]